MGMT-18137: Use _cert-manager_ for Assisted Installer certs
This patch adds the Assisted Installer webhooks, and changes the
manifests so that that they use _cert-manager_ to dynamically generate
the required certificates.

We assume that _cert-manager_ has already been installed with the
default configuration.

We create a `assisted-installer-selfsigned-ca` issuer with the sole
purpose of creating a self signed certificate for our own CA. The result
of this will be our CA private key and certificate inside the
`assisted-installer-ca` secret.  Using that secret we create the
`assisted-installer-ca` issuer, which is then used to generate the all
the certificates that we need. The actual certificates are created using
the _cainjector_ component of _cert-manager_.

The certificates for the ingress are generated creating the required
`Certificate` objects, and adding the following annotation to the
`Ingress` objects:

```yaml
cert-manager.io/issuer: assisted-installer-ca
```

Our webhooks are implemented using the _OpenShift_ generic API service.
One side effect of that is that they use subject access reviews to
verify that the caller has permissions to use the webhook. But plain
vanilla Kubernetes doesn't pass credentials to webhooks. That is why we
need to create an API service for the webhooks, and configure them to
use the `kubernetes` service in the `default` namespace. That way the
webhooks are called with the credentials of the API server, and they can
sucessfully check the caller permissions. That API service gets the
`caBundle` field populated using the `inject-ca-from` annotation.

In addition the patch also needs to do some minor changes to the
existing objects:

1. The `assisted-service` deployment needs to read the ingress
certificate from a secret instead of from a configmap, because
_cert-manager_ stores certificates in secrets.

2. The `assisted-image-service` deployment needs to read the assisted
   service certificate from a secret instead of from a configmap,
   because _cert-manager_ stores certificates in secrets.

3. The names of the files containing the CA certificates need to be
   changed from `ca-bundle.crt` to `ca.crt` because _cert-manager_ uses
   that name in the secrets it generates.

Related: https://issues.redhat.com/browse/MGMT-18137
Related: https://cert-manager.io/docs/concepts/ca-injector
Related: https://github.com/openshift/generic-admission-server
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>